requirements.txt
pandas==2.2.2
numpy==2.3.1
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
python-dotenv==1.2.1
pypdf==6.5.0
orjson==3.11.5
tenacity==9.1.2
pymupdf==1.26.7

openai==2.14.0
transformers==4.57.3
langchain==1.2.0
langchain-openai==1.1.6
langchain-community==0.4.1